requirements.txt
cython==0.29.28
imageio==2.15.0
matplotlib==3.3.4
mxnet==1.9.0
numpy==1.19.5
opencv-python==4.5.5.64
pandas==1.1.5
Pillow==8.4.0
pydensecrf==1.0rc3
pytest==7.0.1
scikit-image==0.17.2
scipy==1.5.4
tensorboard==2.8.0
tensorboardX==2.5
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0
tqdm==4.63.1
